add: enable like and recast reactions when allowReactions boolean is true
